backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "able-chat-backend"
version = "0.1.0"
description = "FastAPI backend for Able Chat"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.0",
  "google-genai>=1.0.0",
  "google-auth>=2.0.0",
  # Database
  "psycopg2-binary==2.9.11",
  "alembic==1.17.1"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "src*"]
